Se agrego la validacion para no borrar marcas o categorias que tengan productos relacionados y sus tests

+++ app/Http/Controllers/Admin/BrandsController.php
@@ -85,6 +85,7 @@ class BrandsController extends Controller
      */
     public function destroy(Brand $brand)
     {
+      if($brand->products->count()) return back()->with(['info' => 'No se puede eliminar la marca si tiene productos asignados.', 'alert' => 'danger']);
       $brand->delete();
 
       return redirect('/administracion/marcas')->with(['info' => "Se ha eliminado la marca {$brand->name}"]);
